+++ datatypes/src/commonMain/kotlin/at/asitplus/crypto/datatypes/asn1/Asn1Encoding.kt
@@ -1,5 +1,3 @@
-@file:OptIn(ExperimentalEncodingApi::class)
-
 package at.asitplus.crypto.datatypes.asn1
 
 import at.asitplus.crypto.datatypes.asn1.BERTags.BIT_STRING
@@ -10,10 +8,7 @@ import at.asitplus.crypto.datatypes.asn1.BERTags.NULL
 import at.asitplus.crypto.datatypes.asn1.BERTags.UTC_TIME
 import at.asitplus.crypto.datatypes.asn1.DERTags.toExplicitTag
 import at.asitplus.crypto.datatypes.io.BitSet
-import io.matthewnelson.encoding.base16.Base16
-import io.matthewnelson.encoding.core.Encoder.Companion.encodeToString
 import kotlinx.datetime.Instant
-import kotlin.io.encoding.ExperimentalEncodingApi
 
 /**
  * Class Providing a DSL for creating arbitrary ASN.1 structures. You will almost certainyl never use it directly, but rather use it as follows:
@@ -69,7 +64,9 @@ class Asn1TreeBuilder {
 
     /**
      * appends a single [Asn1Encodable] to this ASN.1 structure
+     * @throws Asn1Exception in case encoding constrints of children are violated
      */
+    @Throws(Asn1Exception::class)
     fun append(child: Asn1Encodable<*>) = append(child.encodeToTlv())
 
     /**
@@ -133,7 +130,11 @@ class Asn1TreeBuilder {
 
     /**
      * Adds the passed string as PRINTABLE STRING [Asn1Primitive] to this ASN.1 structure
+     *
+     * @throws Asn1Exception if illegal characters are to be encoded into a printable string
      */
+
+    @Throws(Asn1Exception::class)
     fun printableString(value: String) {
         elements += Asn1String.Printable(value).encodeToTlv()
     }
@@ -160,18 +161,19 @@ class Asn1TreeBuilder {
         elements += value.encodeToAsn1GeneralizedTime()
     }
 
+    /**
+     * @throws Asn1Exception if illegal SET or SET OF structures are to be constructed
+     */
+    @Throws(Asn1Exception::class)
     private fun nest(type: CollectionType, init: Asn1TreeBuilder.() -> Unit) {
         val seq = Asn1TreeBuilder()
         seq.init()
-        elements += if (type == CollectionType.SEQUENCE) Asn1Sequence(seq.elements)
-        else if (type == CollectionType.OCTET_STRING) Asn1EncapsulatingOctetString(seq.elements)
-        else Asn1Set(seq.elements.let {
-            if (type == CollectionType.SET) it.sortedBy { it.tag }
-            else {
-                if (it.any { elem -> elem.tag != it.first().tag }) throw IllegalArgumentException("SET_OF must only contain elements of the same tag")
-                it.sortedBy { it.derEncoded.encodeToString(Base16) } //TODO this is inefficient
-            }
-        })
+        elements += when (type) {
+            CollectionType.SEQUENCE -> Asn1Sequence(seq.elements)
+            CollectionType.OCTET_STRING -> Asn1EncapsulatingOctetString(seq.elements)
+            CollectionType.SET -> Asn1Set(seq.elements)
+            CollectionType.SET_OF -> Asn1SetOf(seq.elements)
+        }
     }
 
     /**
@@ -237,7 +239,10 @@ class Asn1TreeBuilder {
      *     }
      *   }
      *  ```
+     *
+     *  @throws Asn1Exception if children have different tags
      */
+    @Throws(Asn1Exception::class)
     fun setOf(init: Asn1TreeBuilder.() -> Unit) = nest(CollectionType.SET_OF, init)
 
     /**
@@ -316,11 +321,36 @@ fun asn1Set(root: Asn1TreeBuilder.() -> Unit): Asn1Set {
  *   printableString("Hello")
  * }
  *  ```
+ *
+ *  @throws Asn1Exception if children of different tags are added
  */
+
+@Throws(Asn1Exception::class)
 fun asn1SetOf(root: Asn1TreeBuilder.() -> Unit): Asn1Set {
     val seq = Asn1TreeBuilder()
     seq.root()
-    return Asn1Set(seq.elements)
+    return Asn1SetOf(seq.elements)
+}
+
+/**
+ * Creates a new EXPLICITLY TAGGED ASN.1 structure as [Asn1Tagged] using [tag].
+ *
+ * * **NOTE:** automatically calls [at.asitplus.crypto.datatypes.asn1.DERTags.toExplicitTag] on [tag]
+ *
+ * Use as follows:
+ *
+ * ```kotlin
+ * asn1Tagged(2u) {
+ *   printableString("World World")
+ *   asn1Null()
+ *   int(1337)
+ * }
+ *  ```
+ */
+fun asn1Tagged(tag: UByte, root: Asn1TreeBuilder.() -> Unit): Asn1Tagged {
+    val seq = Asn1TreeBuilder()
+    seq.root()
+    return Asn1Tagged(tag.toExplicitTag(), seq.elements)
 }
 
 /**