+++ datatypes/src/commonMain/kotlin/at/asitplus/crypto/datatypes/asn1/Asn1Time.kt
@@ -29,6 +29,8 @@ class Asn1Time(val instant: Instant, formatOverride: Format? = null) : Asn1Encod
 
     companion object : Asn1Decodable<Asn1Primitive, Asn1Time> {
         private val THRESHOLD_GENERALIZED_TIME = Instant.parse("2050-01-01T00:00:00Z")
+
+        @Throws(Asn1Exception::class)
         override fun decodeFromTlv(src: Asn1Primitive) =
             Asn1Time(src.readInstant(), if (src.tag == BERTags.UTC_TIME) Format.UTC else Format.GENERALIZED)
     }