+++ datatypes/src/jvmTest/kotlin/PublicKeyTest.kt
@@ -1,7 +1,6 @@
 import at.asitplus.crypto.datatypes.CryptoPublicKey
 import at.asitplus.crypto.datatypes.asn1.Asn1Element
 import at.asitplus.crypto.datatypes.asn1.Asn1Sequence
-import at.asitplus.crypto.datatypes.asn1.ensureSize
 import at.asitplus.crypto.datatypes.asn1.parse
 import at.asitplus.crypto.datatypes.fromJcaKey
 import at.asitplus.crypto.datatypes.getPublicKey
@@ -43,12 +42,12 @@ class PublicKeyTest : FreeSpec({
                 own.shouldNotBeNull()
                 println(Json.encodeToString(own))
                 println(own.iosEncoded.encodeToString(Base16()))
-                println(own.derEncoded.encodeToString(Base16()))
+                println(own.encodeToDer().encodeToString(Base16()))
                 println(own.keyId)
-                own.derEncoded shouldBe pubKey.encoded
+                own.encodeToDer() shouldBe pubKey.encoded
                 CryptoPublicKey.fromKeyId(own.keyId) shouldBe own
                 own.getPublicKey().encoded shouldBe pubKey.encoded
-                CryptoPublicKey.decodeFromTlv(Asn1Element.parse(own.derEncoded) as Asn1Sequence) shouldBe own
+                CryptoPublicKey.decodeFromTlv(Asn1Element.parse(own.encodeToDer()) as Asn1Sequence) shouldBe own
             }
         }
 
@@ -96,8 +95,8 @@ class PublicKeyTest : FreeSpec({
                 val keyBytes = ((ASN1InputStream(pubKey.encoded).readObject()
                     .toASN1Primitive() as ASN1Sequence).elementAt(1) as DERBitString).bytes
                 own.iosEncoded shouldBe keyBytes //PKCS#1
-                own.derEncoded shouldBe pubKey.encoded //PKCS#8
-                CryptoPublicKey.decodeFromTlv(Asn1Element.parse(own.derEncoded) as Asn1Sequence) shouldBe own
+                own.encodeToDer() shouldBe pubKey.encoded //PKCS#8
+                CryptoPublicKey.decodeFromTlv(Asn1Element.parse(own.encodeToDer()) as Asn1Sequence) shouldBe own
                 own.getPublicKey().encoded shouldBe pubKey.encoded
             }
         }